requirements.txt
flask
line-bot-sdk
google-generativeai
pyngrok
Pillow
python-dotenv